+++ CHANGES.rst
@@ -4,11 +4,15 @@ Changelog
 1.0a4 (unreleased)
 ^^^^^^^^^^^^^^^^^^
 
-- Nothing changed yet.
+  * Backend do tile da agenda (closes `#12`_).
+    [rodfersou]
+
+  * Template e CSS do tile agenda (closes `#12`_).
+    [rennan]
 
 
 1.0a3 (2013-11-18)
-^^^^^^^^^^^^^^^^^^^
+^^^^^^^^^^^^^^^^^^
   * Exibe a mensagem de "Sem compromissos oficiais." apenas se
     nao tivermos compromissos e o campo atualizacao nao
     estiver preenchido
@@ -39,14 +43,17 @@ Changelog
     [hvelarde]
 
 1.0a1 (2013-10-29)
-^^^^^^^^^^^^^^^^^^^^^
+^^^^^^^^^^^^^^^^^^
 
   * Estilização de css, templates e portlets laterais (closes `#3`_).
     [felipeduardo]
+
   * Exibicao dos campos na view de agenda diaria (closes `#1`_).
     [dbarbato]
+
   * Versão inicial do pacote.
     [ericof]
 
 .. _`#1`: https://github.com/plonegovbr/brasil.gov.agenda/issues/1
 .. _`#3`: https://github.com/plonegovbr/brasil.gov.agenda/issues/3
+.. _`#12`: https://github.com/plonegovbr/brasil.gov.agenda/issues/12